refactor(types): align SubmissionStatus with API PascalCase spec #219
